flsk_script/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.1
ultralytics==8.0.196 